python-service/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-google-genai
langchain-community
google-generativeai
faiss-cpu
pydantic
pydantic-settings
python-dotenv
httpx
